We can now read the Cache-Control headers
With this new commit we only use the cached response in the headers
say its ok!

This means that we are checking:
- That the requests are for the same Resource. This involves looking at
the Vary headers https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Vary
- They are not too old (stale)

We are using the `http_cache_semantics` crate for this.
Since this requires to full original request in order to determine if its
cachable we store that in memory as well.