We got Filesystem caching to work now!
Before we were holding everything in memory, but we didnt' want to be limited
by the amount of things we can hold in memory. We want to be able to use
our disk to cache more things than can fit in memory.

In theory this will let us deploy to smaller boxes with less memory.