We figured it out! We can use LiteFS Halt correctly now!
There were two big issues that we found today!

First is that we were NOT using the OFD ("Open File Descriptor") locks
correctly. We were instead using the 'full' file locks, which aren't
what LiteFS uses for HALT.

The next issue is that we were closing our File Descriptor before we
actually wrote to the DB. When this File Descriptor was closed, the
OFD lock was released, and the database was NOT writable anymore from
the replica.

The solution is to hold onto the File Descriptor until we're done
writing to the database. This means we open the FD before halting and
use the same instance of the FD to unhalt.